feat(commands): move all commands to slash commands
Parallel to a huge refactor for the reply to interactions, all commands are now processed as slash commands (except for the pin command which became a message context menu command now). This means we do not need messageCreate event anymore (nor do we need the intents related to content).
Changes on the config to remove unnecessary emotes as well.